testes do controle

+++ components/MotorDC/include/MotorDC.h
@@ -49,13 +49,13 @@ public:
   float return_kd();
   void tweak_pid(int variable, float diff);
   double desired_speed_rpm = 0; // velocidade desejada em rpm
-  volatile double current_speed_rpm = 0;
+  volatile int current_speed_rpm = 0;
   volatile double last_error = 0;        // erro anterior para o PID
   volatile double accumulated_error = 0; // erro acumulado para o PID
   bool incrementing = true;
-  volatile int32_t posi = 0;             // posição do motor em ticks do encoder
-  volatile int32_t last_posi = 0;        // posição do motor em ticks do encoder
-  volatile double current_time = 0;
+  volatile int32_t posi = 0;      // posição do motor em ticks do encoder
+  volatile int32_t last_posi = 0; // posição do motor em ticks do encoder
+  volatile long current_time = 0;
   volatile int64_t last_time = 0;
   volatile double angular_position = 0;
   volatile double last_angular_position = 0;
@@ -70,7 +70,7 @@ private:
   float kp;           // valor de kp para o PID
   float ki;           // valor de ki para o PID
   float kd;           // valor de kd para o PID
-  int64_t dt = 0;
+  float dt = 0;
 };
 
 #endif
